% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa.pdb.R
\name{ssa.pdb}
\alias{ssa.pdb}
\title{ggstride}
\usage{
ssa.pdb(pdb)
}
\arguments{
\item{pdb:}{the pdb file including same atoms with the trajectory. bio3d::read.pdb() should be used to read.}
}
\value{
dataframe
}
\description{
ssa.pdb() returns a data frame that contains
residue identifiers (resname_chain_resid; e.g., MET_A_1)
as column names and STRIDE assignments of each residue as
the row
}
\examples{
pdb <- read.pdb("example/wt.pdb")
ssa_df <- ssa.pdb(pdb)

package::ssa(pdb)

}
