% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa_plot.R
\name{ssa_plot}
\alias{ssa_plot}
\title{ggstride}
\usage{
ssa_plot(
  ssa1,
  ssa2,
  name1 = "Wild-type",
  name2,
  resid1 = as.numeric(gsub("\\\\D+", "", colnames(ssa1)[1])),
  resid2 = as.numeric(gsub("\\\\D+", "", colnames(ssa1)[length(colnames(ssa1))])),
  color_number1 = 1,
  color_number2 = 2
)
}
\arguments{
\item{ssa1:}{the ssa data frame calculated for trajectory 1, calculated by ggstride::ssa()}

\item{ssa2:}{the ssa data frame calculated for trajectory 2, calculated by ggstride::ssa()}

\item{name1:}{assign a name for your trajectory 1}

\item{name2:}{assign a name for your trajectory 2}

\item{resid1:}{residue id of the first residue of the region you want to visualize}

\item{resid2:}{residue id of the last residue of the region you want to visualize}

\item{color_number1:}{(between 1 and 7) select a color from jama palette of ggsci for the ssa of trajectory 1}

\item{color_number2:}{(between 1 and 7) select a color from jama palette of ggsci for the ssa of trajectory 2}
}
\value{
plot
}
\description{
ssa_plot() returns a plot that contains percentage of
secondary structures observed along the trajectories per residue by
comparing the two ssa data frames obtained by ssa().
}
\examples{
ssa_plot(ssa1 = ssa_WT, ssa2 = ssa_mutant,
name1 = "Wild-type", name2 = "Leu656Phe",
resid1 = 500, resid2 = 525,
color_number1 = 1, color_number2 = 2)

ggstride::ssa_plot(ssa1, ssa2, name1 = "Wild-type", color_number1 = 1, name2, color_number2)

}
