% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa.R
\name{ssa}
\alias{ssa}
\title{ggstride}
\usage{
ssa(pdb, traj, start = 1, last = nrow(traj))
}
\arguments{
\item{pdb:}{the pdb file including same atoms with the trajectory. bio3d::read.pdb() should be used to read.}

\item{traj:}{the trajectory file, xyz format of bio3d should be used. It could be any type of trajectory supported by bio3d.}

\item{start:}{(default = 1) the number of first frame you want to analyze}

\item{last:}{(default = number of frames in your trajectory) the number of last frame you want to analyze}
}
\value{
dataframe
}
\description{
ssa() returns a data frame that contains
residue identifiers (resname_chain_resid; e.g., MET_A_1)
as column names and STRIDE assignments of each residue as
each row for each frame of the trajectory.
}
\examples{
pdb_WT <- read.pdb("example/wt.pdb")
dcd_WT <- read.dcd(trjfile = "example/wt.dcd")
ssa_WT <- ssa(pdb_WT, dcd_WT)

package::ssa(pdb, traj, start = 1, last = nrow(traj))

}
