% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa_plot.R
\name{ssa_plot}
\alias{ssa_plot}
\title{ggstride}
\usage{
ssa_plot(ssa1, ssa2, name1, name2, color_number1 = 1, color_number2 = 2)
}
\arguments{
\item{ssa1, ssa2, name1, color_number1, name2, color_number2}{}
}
\value{
plot
}
\description{
ssa_plot() returns a plot that contains percentage of
secondary structures observed along the trajectories per residue by
comparing the two ssa data frames obtained by ssa().
}
\examples{
ssa_plot(ssa1 = ssa_WT, ssa2 = ssa_S50P,
name1 = "Wild-type", color_number1 = 1,
name2 = "Ser50Pro", color_number2 = 2)

ggstride::ssa_plot(ssa1, ssa2, name1 = "Wild-type", color_number1 = 1, name2, color_number2)

}
